+++ src/components/Region.js
@@ -9,15 +9,12 @@ export default class Region extends Component {
   }
   
   render() {
-    let reduxGlue = {
-      reduxState: this.props.reduxState,
-      reduxActions: this.props.reduxActions
-
-    }
+    console.log(this);
     return (
       <div className={this.props.className}>
         {this.props.children.map((element,key) => {
-          return React.createElement(Registry.get(element.type), Object.assign(this.props.children[key], reduxGlue))
+          return React.createElement(Registry.get(element.type), getOwnProps(this.props.location.query, key, this.props.children[key], this.props.reduxActions))
+//          return React.createElement(Registry.get(element.type), Object.assign(this.props.children[key], reduxGlue))
         })}
       </div>
     );
