+++ src/utils/utils.js
@@ -50,3 +50,23 @@ export function qFromParams(params) {
 export function paramsFromQ(str) {
   return JSON.parse('{"' + decodeURI(str).replace(/"/g, '\\"').replace(/&/g, '","').replace(/=/g,'":"') + '"}');
 }
+
+/*
+ * Return array of paramaters of format:
+ * @param query {object} the query object from react-router location.query
+ * @param cid {string} unique component ID
+ * @returns {object} the sub-query containing query paramaters keyed to filter ID
+ */
+export function getOwnQueryParams(query, cid) {
+  let ownParams = Object.keys(query)
+    .filter(k => { console.log(k);
+      return (k.indexOf(cid) >= 0)
+    })
+    .map(key => {
+     let fid = key.replace(cid + '_', '');
+     return { fid: fid, value: this.props.appFilterParams[key] }
+    })
+
+  if (ownParams) console.log('OWN PARAMS', ownParams);
+  return ownParams || {};
+}
